demo usage of throttle decorator
