+++ data_structure.py
@@ -748,30 +748,6 @@ def updateNode(self, context):
     """
     self.process_node(context)
 
-def throttled_nodetree_and_auto_update(func):
-    """
-    use as a decorator
-
-        @throttled_nodetree_and_auto_update
-        def mode_update(self, context):
-            ...
-
-    When a node has changed, like a mode-change leading to a socket change (remove, add, hide...)
-    Blender will trigger nodetree.update. We want to ignore this trigger-event, and we do so by
-    - first throttling the update system. 
-    - then We execute the code that makes changes to the node/nodetree
-    - then we end the throttle-state
-    - we are then ready to process
-
-    """
-    def wrapper_update(self, context):
-        with self.sv_throttle_tree_update():
-            func(self, context)
-        self.process_node(context)
-
-    return wrapper_update
-
-
 
 ##############################################################
 ##############################################################
