+++ nodes/CAD/crop_mesh_2d.py
@@ -8,7 +8,7 @@
 
 import bpy
 
-from sverchok.node_tree import SverchCustomTreeNode, throttle_tree_update
+from sverchok.node_tree import SverchCustomTreeNode
 from sverchok.data_structure import updateNode
 from sverchok.utils.geom_2d.merge_mesh import crop_mesh, crop_edges
 
@@ -24,15 +24,15 @@ class SvCropMesh2D(bpy.types.Node, SverchCustomTreeNode):
     bl_label = 'Crop mesh 2D'
     bl_icon = 'MOD_BOOLEAN'
 
+    @throttled
     def update_sockets(self, context):
-        with throttle_tree_update(self):
-            [self.outputs.remove(sock) for sock in self.outputs[2:]]
-            if self.face_index and self.input_mode == 'faces':
-                self.outputs.new('SvStringsSocket', 'Face index')
-            if self.inputs[1].name != self.input_mode:
-                self.inputs[1].name = self.input_mode.title()
-                self.outputs[1].name = self.input_mode.title()
-        updateNode(self, context)
+        [self.outputs.remove(sock) for sock in self.outputs[2:]]
+        if self.face_index and self.input_mode == 'faces':
+            self.outputs.new('SvStringsSocket', 'Face index')
+        if self.inputs[1].name != self.input_mode:
+            self.inputs[1].name = self.input_mode.title()
+            self.outputs[1].name = self.input_mode.title()
+
 
     mode_items = [('inner', 'Inner', 'Fit mesh', 'SELECT_INTERSECT', 0),
                   ('outer', 'Outer', 'Make hole', 'SELECT_SUBTRACT', 1)]
