+++ nodes/CAD/merge_mesh_2d_lite.py
@@ -8,7 +8,7 @@
 
 import bpy
 
-from sverchok.node_tree import SverchCustomTreeNode, throttle_tree_update
+from sverchok.node_tree import SverchCustomTreeNode
 from sverchok.data_structure import updateNode
 from sverchok.utils.geom_2d.merge_mesh import merge_mesh_light
 
@@ -24,18 +24,18 @@ class SvMergeMesh2DLite(bpy.types.Node, SverchCustomTreeNode):
     bl_label = 'Merge mesh 2D lite'
     bl_icon = 'AUTOMERGE_ON'
 
+    @throttled
     def update_sockets(self, context):
-        with throttle_tree_update(self):
-            links = {sock.name: [link.to_socket for link in sock.links] for sock in self.outputs}
-            [self.outputs.remove(sock) for sock in self.outputs[2:]]
-            new_socks = []
-            if self.face_index:
-                new_socks.append(self.outputs.new('SvStringsSocket', 'Face index'))
-            if self.overlap_number:
-                new_socks.append(self.outputs.new('SvStringsSocket', 'Overlap number'))
-            [[self.id_data.links.new(sock, link) for link in links[sock.name]]
-                                                 for sock in new_socks if sock.name in links]
-        updateNode(self, context)
+        links = {sock.name: [link.to_socket for link in sock.links] for sock in self.outputs}
+        [self.outputs.remove(sock) for sock in self.outputs[2:]]
+        new_socks = []
+        if self.face_index:
+            new_socks.append(self.outputs.new('SvStringsSocket', 'Face index'))
+        if self.overlap_number:
+            new_socks.append(self.outputs.new('SvStringsSocket', 'Overlap number'))
+        [[self.id_data.links.new(sock, link) for link in links[sock.name]]
+                                             for sock in new_socks if sock.name in links]
+
 
     face_index: bpy.props.BoolProperty(name="Show face mask", update=update_sockets,
                                        description="Show output socket of index face mask")
